+++ packages/server/customer-os-neo4j-repository/entity/phone_number.go
@@ -1,18 +1,16 @@
 package entity
 
 import (
-	neo4jentity "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/entity"
 	"time"
 )
 
-// Deprecated
 type PhoneNumberEntity struct {
 	Id             string
 	E164           string
 	Validated      *bool
 	RawPhoneNumber string
-	Source         neo4jentity.DataSource
-	SourceOfTruth  neo4jentity.DataSource
+	Source         DataSource
+	SourceOfTruth  DataSource
 	AppSource      string
 	CreatedAt      time.Time
 	UpdatedAt      time.Time
