+++ packages/server/customer-os-neo4j-repository/test/neo4j_data_helper.go
@@ -630,6 +630,35 @@ func CreateLocation(ctx context.Context, driver *neo4j.DriverWithContext, tenant
 	return locationId
 }
 
+func CreatePhoneNumber(ctx context.Context, driver *neo4j.DriverWithContext, tenant string, phoneNumber entity.PhoneNumberEntity) string {
+	phoneNumberId := utils.NewUUIDIfEmpty(phoneNumber.Id)
+	query := fmt.Sprintf(`MATCH (t:Tenant {name: $tenant})
+			  MERGE (t)<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(i:PhoneNumber {id:$id})
+				SET i:PhoneNumber_%s,
+					i.e164=$e164,
+					i.validated=$validated,
+					i.rawPhoneNumber=$rawPhoneNumber,
+					i.source=$source,
+					i.sourceOfTruth=$sourceOfTruth,
+					i.appSource=$appSource,
+					i.createdAt=$createdAt,
+					i.updatedAt=$updatedAt`, tenant)
+
+	ExecuteWriteQuery(ctx, driver, query, map[string]any{
+		"tenant":         tenant,
+		"id":             phoneNumberId,
+		"e164":           phoneNumber.E164,
+		"validated":      phoneNumber.Validated,
+		"rawPhoneNumber": phoneNumber.RawPhoneNumber,
+		"source":         phoneNumber.Source,
+		"sourceOfTruth":  phoneNumber.SourceOfTruth,
+		"appSource":      phoneNumber.AppSource,
+		"createdAt":      phoneNumber.CreatedAt,
+		"updatedAt":      phoneNumber.UpdatedAt,
+	})
+	return phoneNumberId
+}
+
 func CreateContractForOrganization(ctx context.Context, driver *neo4j.DriverWithContext, tenant, organizationId string, contract entity.ContractEntity) string {
 	contractId := utils.NewUUIDIfEmpty(contract.Id)
 	query := fmt.Sprintf(`MATCH (t:Tenant {name:$tenant}), (o:Organization {id:$organizationId})
