--- packages/server/events-processing-platform-subscribers/graph_db/entity/interaction_session_participant_interface.go
@@ -1,13 +1,0 @@
-package entity
-
-type InteractionSessionParticipantDetails struct {
-	Type string
-}
-
-type InteractionSessionParticipant interface {
-	IsInteractionSessionParticipant()
-	ParticipantLabel() string
-	GetDataloaderKey() string
-}
-
-type InteractionSessionParticipants []InteractionSessionParticipant
