--- packages/server/events-processing-platform-subscribers/graph_db/entity/meeting_participant_interface.go
@@ -1,9 +1,0 @@
-package entity
-
-type MeetingParticipant interface {
-	IsMeetingParticipant()
-	MeetingParticipantLabel() string
-	GetDataloaderKey() string
-}
-
-type MeetingParticipants []MeetingParticipant
