--- packages/server/events-processing-platform-subscribers/graph_db/entity/mentioned_entity_interface.go
@@ -1,10 +1,0 @@
-package entity
-
-// Deprecated
-type MentionedEntity interface {
-	IsMentionedEntity()
-	MentionedEntityLabel() string
-	GetDataloaderKey() string
-}
-
-type MentionedEntities []MentionedEntity
