--- packages/server/events-processing-platform-subscribers/graph_db/entity/noted_entity_interface.go
@@ -1,9 +1,0 @@
-package entity
-
-type NotedEntity interface {
-	IsNotedEntity()
-	NotedEntityLabel() string
-	GetDataloaderKey() string
-}
-
-type NotedEntities []NotedEntity
