+++ packages/server/events-processing-platform-subscribers/subscriptions/graph/user_event_handlers_it_test.go
@@ -434,7 +434,7 @@ func TestGraphUserEventHandler_OnPhoneNumberLinkedToUser(t *testing.T) {
 	validated := false
 	e164 := "+0123456789"
 
-	phoneNumberId := neo4jt.CreatePhoneNumber(ctx, testDatabase.Driver, tenantName, entity.PhoneNumberEntity{
+	phoneNumberId := neo4jtest.CreatePhoneNumber(ctx, testDatabase.Driver, tenantName, neo4jentity.PhoneNumberEntity{
 		E164:           e164,
 		Validated:      &validated,
 		RawPhoneNumber: e164,
