+++ packages/server/events-processing-platform-subscribers/test/neo4j/neo4j_data_helper.go
@@ -108,35 +108,6 @@ func CreateComment(ctx context.Context, driver *neo4j.DriverWithContext, tenant 
 	return commentId
 }
 
-func CreatePhoneNumber(ctx context.Context, driver *neo4j.DriverWithContext, tenant string, phoneNumber entity.PhoneNumberEntity) string {
-	phoneNumberId := utils.NewUUIDIfEmpty(phoneNumber.Id)
-	query := fmt.Sprintf(`MATCH (t:Tenant {name: $tenant})
-			  MERGE (t)<-[:PHONE_NUMBER_BELONGS_TO_TENANT]-(i:PhoneNumber {id:$id})
-				SET i:PhoneNumber_%s,
-					i.e164=$e164,
-					i.validated=$validated,
-					i.rawPhoneNumber=$rawPhoneNumber,
-					i.source=$source,
-					i.sourceOfTruth=$sourceOfTruth,
-					i.appSource=$appSource,
-					i.createdAt=$createdAt,
-					i.updatedAt=$updatedAt`, tenant)
-
-	neo4jtest.ExecuteWriteQuery(ctx, driver, query, map[string]any{
-		"tenant":         tenant,
-		"id":             phoneNumberId,
-		"e164":           phoneNumber.E164,
-		"validated":      phoneNumber.Validated,
-		"rawPhoneNumber": phoneNumber.RawPhoneNumber,
-		"source":         phoneNumber.Source,
-		"sourceOfTruth":  phoneNumber.SourceOfTruth,
-		"appSource":      phoneNumber.AppSource,
-		"createdAt":      phoneNumber.CreatedAt,
-		"updatedAt":      phoneNumber.UpdatedAt,
-	})
-	return phoneNumberId
-}
-
 func LinkIssueReportedBy(ctx context.Context, driver *neo4j.DriverWithContext, issueId, entityId string) {
 	query := `MATCH (e {id:$entityId})
 				MATCH (i:Issue {id:$issueId})
